Fix favorites module qa-functional tests.
