+++ favorites/test/qa-functional/src/org/netbeans/modules/favorites/HomeFolderTest.java
@@ -1,0 +1,62 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package org.netbeans.modules.favorites;
+
+import java.io.File;
+import junit.framework.Test;
+import org.netbeans.jellytools.FavoritesOperator;
+import org.netbeans.jellytools.JellyTestCase;
+import org.netbeans.jellytools.actions.FavoritesAction;
+import org.netbeans.jellytools.nodes.Node;
+import org.netbeans.junit.NbModuleSuite;
+
+/**
+ * Basic favorites test.
+ *
+ * @author Tomas.Musil@sun.com
+ */
+public class HomeFolderTest extends JellyTestCase {
+
+    /** 
+     * Need to be defined because of JUnit
+     * @param name 
+     */
+    public HomeFolderTest(String name) {
+        super(name);
+    }
+
+    public static Test suite() {
+        return NbModuleSuite.create(HomeFolderTest.class, ".*", ".*");
+    }
+
+    public @Override void setUp() {
+        JellyTestCase.closeAllModal();
+    }
+
+    public void testHomeFolder(){
+        // checking if $HOME is by default present in favorites
+        new FavoritesAction().performShortcut();
+        FavoritesOperator fo = FavoritesOperator.invoke();
+        File home = new File(System.getProperty("user.home"));
+        Node nodeHome = new Node(fo.tree(),home.getName());
+        nodeHome.expand();
+        nodeHome.collapse();
+    }
+
+}
